torn

+++ magic.js
@@ -2,6 +2,8 @@
 d = document.createElement("canvas");
 e = d.getContext("2d");
 
+// c.globalCompositeOperation = "lighten";
+
 // debug = document.createElement("div");
 // b.appendChild(debug)
 // b.appendChild(d);
@@ -13,38 +15,41 @@ var size = a.width = a.height = 512,
 	beginWarp = false, 
 	third = M.PI / 3, 
 	parts = [], 
-	dead = [],
-	i;
+	dead = [];
 
 d.width = d.height = size * 2;
 
-var r = (v) => { return ~~(M.random() * v || 1); };
+var r = v => ~~(M.random() * v);
 
-var sixit = (v) => { return (v + 6) % 6 }; // clamp v between 0 and 6
+var sixit = v => (v + 6) % 6; // clamp v between 0 and 6
 
-
-var create = (o, x, y, d, mod) => {
+var create = (parent, d, mod) => {
 	var p;
-	x = x || size;
-	y = y || size;
+	var x = parent ? parent.x : size;
+	var y = parent ? parent.y : size;
 	mod = mod || 1;
+	var o = !parent;
+	// con.log(o)
 	// con.log(x, y);
+
+
 
 	// e.fillStyle = "blue"
 	// e.fillRect(x - 5, y - 5, 10, 10);//p.s, p.s);
 
 	
-	if (dead.length) {
+	if (parent && dead.length) {
 		p = dead.splice(-1)[0];
 		p.d = d;
 		p.x = p.lx = x;
 		p.y = p.ly = y;
 		p.o = false;
 		p.pos = 0;
-		p.dying = 1;
+		p.dying = r(4) + 1;
 		p.alive = true;
+		p.colour = parent.colour;
 
-		p.setColour();
+		//p.setColour();
 		
 	} else {
 
@@ -64,26 +69,28 @@ var create = (o, x, y, d, mod) => {
 			life: 0,
 			d: d || r(3) * 2, // 0, 2 or 4
 			mod,
-			s: 2 / 1 * mod, //1 / M.pow(2, r(2)),
-			v: 1 / 32 / mod, // / M.pow(2, r(4)),
-			setColour: (p2) => {
-				p.colour = "hsla(" + (p.life++ * 20) + ",50%," + (20 + p.s * 20) +"%,0.7)";
-				// p.colour = "hsla(" + r(360) + ",70%,50%,1)";
+			s: 2 * mod,
+			// s: 1 / M.pow(2, r(2)),
+			v: 1 / 32,// / mod, // / M.pow(2, r(4)),
+			// setColour: () => {
+				// p.colour = "hsla(" + (p.life++ * 20) + ",50%," + (20 + p.s * 20) +"%,0.7)";
+			colour: parent ? parent.colour : "hsla(" + parts.length * 80 + ",50%,50%,0.6)",
+				// p.colour = parent ? parent.colour : "hsla(" + r(360) + ",50%,50%,1)";
 				// con.log(p.colour)
-			},
+			// },
 			kill: () => {
-				p.x = 0;
-				p.y = 0;
+				// p.x = 0;
+				// p.y = 0;
 				p.alive = false;
 				dead.push(p);
 			},
 			m: () => {
 
-				// if (p.alive == false) {
-				// 	e.fillStyle = p.colour;
-				// 	e.fillText(p.index, p.index * 5, 10);
-				// 	return;//con.log("returning")
-				// }
+				if (p.alive == false) {
+					// e.fillStyle = p.colour;
+					// e.fillText(p.index, p.index * 5, 10);
+					return;//con.log("returning")
+				}
 
 				p.dying *= p.o ? 1 : 0.95;
 
@@ -96,7 +103,11 @@ var create = (o, x, y, d, mod) => {
 				if (p.pos==1) {
 
 					if (p.x < 0 || p.y > size * 2 || p.y < 0 || p.y > size * 2) {
-						p.kill();
+						if (p.o) {
+							create();
+						} else {
+							p.kill();
+						}
 						return;
 					}
 
@@ -105,21 +116,21 @@ var create = (o, x, y, d, mod) => {
 					// p.setColour();
 					p.d += sixit(r(2) * 2 - 1); // add -1 or 1 to new dir.
 
-					if (p.mod > 1/4 && r(10) > 2) { // duplicate at current position
+					if (p.mod > 1 / 4 && r(10) > 2) { // duplicate at current position
 						
 						// var newDir = r(2) * 4 - 2; // + -2 or 2
 						// newDir = p.d + newDir; // make sure clone has new direction
 						// newDir = (newDir + 6) % 6; // clamp to positives: 0 > 5
-
-						for (var i = 0; i < 2; i++) {
-							
+						for (var i = r(6); i--;) {
+							// con.log(i)
 							
 							var newDir = sixit(p.d + r(2) * 4 - 2); // same as above 3 lines
 							
 							// e.fillStyle = "red"
 							// e.fillRect(p.x - 10, p.y - 10, 20, 20);//p.s, p.s);
 
-							create(false, p.x, p.y, newDir, p.mod / 2);
+							// create(false, p.x, p.y, newDir, p.mod / 2);
+							create(p, newDir, p.mod / 2);
 							// debugger;
 						}
 					}
@@ -156,8 +167,8 @@ var create = (o, x, y, d, mod) => {
 			}
 		}
 
-		p.setColour();
-		p.index = parts.length;
+		// p.setColour();
+		// p.index = parts.length;
 		parts.push(p);
 
 
@@ -169,9 +180,9 @@ var render = (t) =>{
 	// debug.innerHTML = parts.map(p=>Math.round(p.x)); 
 
 	if ((M.floor(t / 1000) + 1) % 3 == 0) {
-		if (beginWarp == false) { // warp has just begun! fuck yeah.
+		if (!beginWarp) { // warp has just begun! fuck yeah.
 			beginWarp = true;
-			scaleTarget = M.random() + .7;
+			scaleTarget = M.random() + .7; // Math.sqrt(2) / 2 is min scale
 		}
 	} else {
 		beginWarp = false;
@@ -185,18 +196,18 @@ var render = (t) =>{
 	c.rotate(t * 0.0001); // arbitrary divisor
 	c.translate(-size, -size);
 
-	e.fillStyle = "rgba(0,0,0,0.03)";
+	e.fillStyle = "rgba(0,0,0,0.04)";
 	e.fillRect(0, 0, size * 2, size * 2);
 
 
 	// if (!warpMode)
-	for(i=0;i<parts.length;i++) parts[i].m();
+	for(var i=0;i<parts.length;i++) parts[i].m();
 
 	c.drawImage(d, 0, 0);
 	c.restore();
 	requestAnimationFrame(render);
 };
 
-while(parts.length < 4) create(true);
+while(parts.length < 4) create();
 
 render(0);
